merge Haskell files again
rationale: now refactor.ts on victortaelin/ai-cripts uses the old block
based approach for omitting and patching code, meaning we don't rely on
imports for that anymore. that allows us to have one file again. having
one file is favorable because it allows us to undo a bad AI job by just
using the editor's undo option, so we don't need to involve git or any
other source control. this is a very convenient & efficient sweetspot