fix: the following hvm program would hit snf's TODO for a stuck Dup:
```
@main =
  λa.
    !b&T = a; &T{
      !c&U = b₀; &U{
        !c&V = c₀; &V{c₀,&{}},
        &X{&{},&{}}
      }
      ,
      λ{
        #Z: !d&X = b₁; &X{d₀,&{}};
        #S: &Y{&{},&{}};
      }
    }

```

This Dup comes from the `wnf` of a Cop, which entered the fallback case `return (Dup k l v (Cop s k))`.
This is the production of a Dup node whose substitution was originally already in the dup_map, and
since it has been removed a few lines earlier, although the Dup exists, it's not available for a substitution anymore.

Further inspection shows that, at the time of the TODO error, this newly produced produced Dup is a
subterm of the value of a Dup that lives in the `dup_map`, and since we have lazy evaluation,
this subterm is never reached, and this Dup never becomes a substitution.

We remedy this by, instead of returning (Dup ...), putting the Dup back into the `dup_map` and returning
just (Cop s k), which is the continuation of this Dup, i.e. it's wnf.